requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
python-dotenv==1.0.0
gunicorn==21.2.0
pywhatkit==5.4
APScheduler==3.10.4
requests==2.31.0